requirements.txt
streamlit
pymupdf
jinja2
weasyprint
python-dotenv
openai
google-genai
argostranslate
pytesseract
Pillow
